Rename wait_async to wait_asynch
